refactor: add undocumented root_package attribute to npm_traslate_lock
