+++ e2e/npm_link_package-esm/BUILD.bazel
@@ -1,6 +1,6 @@
 load("@npm//:defs.bzl", "npm_link_all_packages")
 load("@aspect_rules_js//npm:defs.bzl", "npm_link_package")
-load("@rules_foo_npm//foo:@aspect-test/a/package_json.bzl", aspect_test_a_bin = "bin")
+load("@foo//:@aspect-test/a/package_json.bzl", aspect_test_a_bin = "bin")
 
 npm_link_all_packages(name = "node_modules")
