--- e2e/rules_foo/BUILD.bazel
@@ -1,8 +1,0 @@
-load("@aspect_bazel_lib//lib:write_source_files.bzl", "write_source_files")
-
-write_source_files(
-    name = "write_npm_repositories",
-    files = {
-        "npm_repositories.bzl": "@rules_foo_npm//:repositories.bzl",
-    },
-)
