+++ e2e/rules_foo/WORKSPACE
@@ -17,6 +17,10 @@ nodejs_register_toolchains(
     node_version = "16.9.0",
 )
 
-load("@rules_foo//:repositories.bzl", rules_foo_repositories = "repositories")
+load("@rules_foo//foo:repositories.bzl", "foo_repositories")
 
-rules_foo_repositories()
+foo_repositories(name = "foo")
+
+load("@foo//:npm_repositories.bzl", foo_npm_repositories = "npm_repositories")
+
+foo_npm_repositories()
